Barebones definitions of all the types defined in the OpenAPI spec.
This includes every type that will be necessary, with the exception of implementing the ranges and other fields that are limited in one way or another.
I also added the descriptions from the OpenAPI spec (where there were any), to help with understanding the types and fields when developing in the future.